requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
drf-nested-routers==0.92.5
requests==2.25.1
